+++ src/app/public/app-config-host.spec.ts
@@ -1,6 +1,6 @@
 import {
   SkyAppConfigHost
-} from './host';
+} from './app-config-host';
 
 describe('SkyAppConfigHost', () => {
